+++ src/main/model-legacy-extension-api.ts
@@ -2,7 +2,7 @@ import type { LegacyPrepareContextApi, PrepareContextApiWrapper, Model, Query, P
 import type { ValueAnnotation, ValueAnnotationContext } from '../views/struct/render-annotations.js';
 import type { ObjectMarkerConfig } from '../core/object-marker.js';
 import { ObjectMarkerManager } from '../core/object-marker.js';
-import { hasOwn } from '../core/utils/object-utils.js';
+import modelCommonJoraMethods from './model-common-jora-methods.js';
 import jora from 'jora';
 
 export function createLegacyExtensionApi(host: Model, options?: SetDataOptions): PrepareContextApiWrapper {
@@ -31,8 +31,8 @@ export function createLegacyExtensionApi(host: Model, options?: SetDataOptions):
         }
     };
     let queryCustomMethods = {
+        ...modelCommonJoraMethods,
         query: host.query.bind(host),
-        overrideProps,
         pageLink: (pageRef: PageRef, pageId?: string, pageParams?: PageParams, pageAnchor?: PageAnchor) =>
             host.encodePageHash(pageId, pageRef, pageParams, pageAnchor),
         marker: lookupObjectMarker,
@@ -107,22 +107,6 @@ export function createLegacyExtensionApi(host: Model, options?: SetDataOptions):
 
     function lookupObjectMarkerAll(value: unknown) {
         return objectMarkers.lookupAll(value);
-    }
-
-    function overrideProps(current: any, props = this.context.props) {
-        if (!props) {
-            return current;
-        }
-
-        const result = { ...current };
-
-        for (const key of Object.keys(result)) {
-            if (hasOwn(props, key)) {
-                result[key] = props[key];
-            }
-        }
-
-        return result;
     }
 
     function addValueAnnotation(query: Query, options: object | boolean = false) {